# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Focus Communication System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User management
  - name: contacts
    description: Contact management
  - name: sessions
    description: Focus session management
  - name: messages
    description: Message management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithToken"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithToken"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /contacts:
    get:
      operationId: getContacts
      tags: [contacts]
      summary: Get all contacts for current user
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactWithUser"
    post:
      operationId: addContact
      tags: [contacts]
      summary: Add a contact
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddContactBody"
      responses:
        "201":
          description: Contact added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactWithUser"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already a contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contacts/{contactId}:
    delete:
      operationId: removeContact
      tags: [contacts]
      summary: Remove a contact
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: path
          name: contactId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Contact removed
        "404":
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions:
    get:
      operationId: getSessions
      tags: [sessions]
      summary: Get all sessions for current user
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, completed]
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionWithParticipants"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new focus session
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWithParticipants"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already has an active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{sessionId}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get a specific session
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWithParticipants"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSession
      tags: [sessions]
      summary: Update a session (e.g., end it)
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionBody"
      responses:
        "200":
          description: Session updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWithParticipants"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{sessionId}/invite:
    post:
      operationId: inviteToSession
      tags: [sessions]
      summary: Invite a contact to a session
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteToSessionBody"
      responses:
        "200":
          description: Participant added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWithParticipants"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{sessionId}/join:
    post:
      operationId: joinSession
      tags: [sessions]
      summary: Join a session you were invited to
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Joined session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWithParticipants"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{sessionId}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a session
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: before
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in a session
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{sessionId}/messages/poll:
    get:
      operationId: pollMessages
      tags: [messages]
      summary: Poll for new messages since a timestamp
      parameters:
        - in: header
          name: x-user-id
          required: true
          schema:
            type: string
        - in: path
          name: sessionId
          required: true
          schema:
            type: integer
        - in: query
          name: since
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: New messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - createdAt

    UserWithToken:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        token:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - name
        - token
        - createdAt

    RegisterUserBody:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
      required:
        - username
        - name
        - password

    LoginUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    ContactWithUser:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        contactUser:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - contactUser
        - createdAt

    AddContactBody:
      type: object
      properties:
        contactUserId:
          type: integer
      required:
        - contactUserId

    SessionParticipant:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        sessionId:
          type: integer
        status:
          type: string
          enum: [invited, joined]
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - userId
        - sessionId
        - status
        - user

    Session:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        creatorId:
          type: integer
        status:
          type: string
          enum: [active, completed]
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - title
        - creatorId
        - status
        - createdAt

    SessionWithParticipants:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        creatorId:
          type: integer
        status:
          type: string
          enum: [active, completed]
        participants:
          type: array
          items:
            $ref: "#/components/schemas/SessionParticipant"
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - title
        - creatorId
        - status
        - participants
        - createdAt

    CreateSessionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title

    UpdateSessionBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed]
        title:
          type: string
        description:
          type: string

    InviteToSessionBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId

    Message:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        status:
          type: string
          enum: [sent, delivered, read]
        createdAt:
          type: string
          format: date-time
        sender:
          $ref: "#/components/schemas/User"
      required:
        - id
        - sessionId
        - senderId
        - content
        - status
        - createdAt
        - sender

    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content