# lib/api-zod/src/generated/types/createSessionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Focus Communication System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSessionBody {
  title: string;
  description?: string;
}