# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Focus Communication System API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionStatus } from "./sessionStatus";

export interface Session {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  creatorId: number;
  status: SessionStatus;
  createdAt: Date;
  /** @nullable */
  endedAt?: Date | null;
}