# lib/api-zod/src/generated/types/sessionWithParticipants.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Focus Communication System API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionParticipant } from "./sessionParticipant";
import type { SessionWithParticipantsStatus } from "./sessionWithParticipantsStatus";

export interface SessionWithParticipants {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  creatorId: number;
  status: SessionWithParticipantsStatus;
  participants: SessionParticipant[];
  createdAt: Date;
  /** @nullable */
  endedAt?: Date | null;
}